Enable packet proxy on ARM architectures
I couldn't find any explanation as to why the packet proxy isn't
enabled on ARM targets.

I've been successfully building/using it on a Raspberry Pi for the
last couple weeks. Is there a reason it shouldn't be enabled?

(The diff makes it look complicated, but I just copied
`packet_proxy_linux_amd64.go` to `packet_proxy_linux.go` and
then deleted `packet_proxy_linux_amd64.go`.)
